RG-2542 API / internal separation
