+++ src/table/table.tsx
@@ -1,296 +1,209 @@
-import Table from './table-component';
+import React, {type ComponentPropsWithRef, useCallback, useRef, useState} from 'react';
+import classNames from 'classnames';
 
-import type {AriaAttributes, ReactNode, RefObject} from 'react';
+import {IntersectionObserverContext} from '../global/intersection-observer-context';
+import {SpacerRow, useTableVirtualize as useVirtualItems} from './internal/virtual-items';
+import {DefaultItemRenderer, keyboardFocusableAttrName} from './default-item-renderer';
+import {
+  type AnimatedColumn,
+  AnimatedColumnContext,
+  CollapseItemIntoSpacerContext,
+  defaultRowHeight,
+  TablePropsContext,
+} from './table-const';
+import {focusWithTemporaryTabIndex} from '../global/focus-with-temporary-tabindex';
+import {useAnimatedColumn} from './internal/animated-column';
+import {useComposedRef} from '../global/compose-refs';
+import {TableHeader} from './internal/table-header';
 
-export default Table;
+import type {TableProps} from './table-props';
 
-export interface TableProps<T> {
-  /**
-   * The data items to render. `null` and `undefined` as an item is not supported.
-   * Referentially same items are not supported either.
-   */
-  data: T[];
-
-  /**
-   * Columns specification.
-   */
-  columns: Column<T>[];
-
-  /**
-   * Used to render a row key, e.g. `<tr key={getKey(item, index)}>`.
-   */
-  getKey: (item: T, index: number, items: T[]) => React.Key;
-
-  /**
-   * If true, the table header will not be rendered.
-   */
-  noHeader?: boolean;
-
-  /**
-   * Called when the client moves a row by dragging it.
-   */
-  onItemMove?: (item: T, fromIndex: number, toIndex: number, items: T[]) => void;
-
-  /**
-   * Called when the client clicks on SortButton in a column header.
-   */
-  onSort?: (columnIndex: number, newOrder: SortOrder, columns: Column<T>[]) => void;
-
-  /**
-   * Called when the client clicks on a column delete button in the header.
-   */
-  onColumnDelete?: (columnIndex: number, columns: Column<T>[]) => void;
-
-  /**
-   * Called when the user reorders columns by dragging a column.
-   * The `insertionIndex` parameter represents an insertion position in the original,
-   * unchanged `columns` array before the column is removed.
-   *
-   * One possible implementation is:
-   *
-   * ```ts
-   * const [moved] = columns.splice(fromIndex, 1);
-   * columns.splice(fromIndex < insertionIndex ? insertionIndex - 1 : insertionIndex, 0, moved);
-   * ```
-   *
-   * The callback is not called when the reorder operation would not change the
-   * column order, i.e. when
-   * `insertionIndex === fromIndex || insertionIndex === fromIndex + 1`.
-   */
-  onColumnReorder?: (fromIndex: number, insertionIndex: number, columns: Column<T>[]) => void;
-
-  /**
-   * By default, when a column is reordered, the moved column is highlighted
-   * with a temporary background color. Set `true` to disable this animation.
-   */
-  noColumnReorderAnimation?: boolean;
-
-  /**
-   * Implement to specify item-scoped props like `clickable`, handlers like `onClick`,
-   * a custom `className`, a `ref` etc. like this:
-   *
-   * ```tsx
-   * <Table
-   *   renderItem={(item, index) => (
-   *     <DefaultItemRenderer
-   *       index={index}
-   *       className='my-row'
-   *     />
-   *   )}
-   * />
-   * ```
-   *
-   * More examples: see `DefaultItemRenderer` in `default-item-renderer.tsx`, and the stories.
-   *
-   * You can also use `renderItem` to render custom row(s) with your own structure and styles.
-   *
-   * In your custom implementation, use `TableRow` and `TableCell` base components to apply the
-   * standard classnames, but beware that `tdClassName` won't be applied.
-   *
-   * You can also use `DefaultItemRenderer` in combination with your custom rows.
-   */
-  renderItem?: (item: T, index: number, items: T[]) => ReactNode;
-
-  /**
-   * Only renders rows near the viewport.
-   *
-   * Rows may transition between two states:
-   * - materialized: rendered as actual table rows. This happens when the corresponding
-   *   spacer approaches the viewport, as specified by `lookaheadPx`.
-   * - virtualized: replaced with spacer rows of the same height. This happens when the row
-   *   moves sufficiently far from the viewport, as specified by `retentionMarginPx`.
-   */
-  virtualizeRows?: boolean;
-
-  /**
-   * Used with `virtualizeRows` as a source of scroll events, as a target for ResizeObserver,
-   * and as the root for IntersectionObserver. Required when the scrollable container is not
-   * the whole document.
-   *
-   * If not set:
-   * - scroll listener is attached to `window`
-   * - ResizeObserver observes `document.body`
-   * - IntersectionObserver has no root (i.e. the viewport is used)
-   */
-  scrollerRef?: RefObject<HTMLElement | null>;
-
-  /**
-   * Used with `virtualizeRows` to estimate the height of items that have not been rendered yet.
-   * The function should be fast and side-effect free. Do not measure the DOM here.
-   * Once a row is rendered, its actual height will be measured and used instead of this estimate.
-   *
-   * Note the effects of imprecise estimates:
-   * - When the height is underestimated, the table may materialize more rows than specified by `lookaheadPx`.
-   *   If the resulting rows extend beyond `retentionMarginPx`, they will be virtualized again.
-   *   If this causes relayout flickering, increase `retentionMarginPx`.
-   * - When the height is overestimated, the table may materialize fewer rows than specified by `lookaheadPx`,
-   *   which may leave a spacer partially visible. To avoid this, increase `lookaheadPx` (and
-   *   `retentionMarginPx` accordingly, since it should be greater than `lookaheadPx`).
-   *
-   * Default: 37px = 16px padding + 20px line height + 1px border.
-   */
-  estimateHeight?: (item: T, index: number, items: T[]) => number;
-
-  /**
-   * When using `virtualizeRows`, the number of pixels above and below the viewport
-   * to materialize in advance.
-   *
-   * Increase this value if blank space becomes visible during fast scrolling.
-   *
-   * Default: 400px.
-   */
-  lookaheadPx?: number;
-
-  /**
-   * Used with `virtualizeRows`. Additional margin around the viewport before
-   * materialized rows become eligible for virtualization.
-   *
-   * Increasing this value reduces row churn when heights are underestimated.
-   * In that case, the table may materialize more rows than needed and then immediately
-   * virtualize them again. A larger margin keeps such rows rendered for longer,
-   * at the cost of rendering more rows overall.
-   *
-   * This value should be greater than `lookaheadPx`.
-   * Increase it if you notice table relayouts during initial render or scrolling.
-   *
-   * Default: 450px.
-   */
-  retentionMarginPx?: number;
-
-  /**
-   * When using `virtualizeRows`, ignore scroll and resize position changes
-   * smaller than this value.
-   *
-   * Measurement inaccuracies and rounding artifacts may slightly change the
-   * table layout during materialization and virtualization. With scroll
-   * anchoring enabled (the default browser behavior), the browser may then
-   * adjust the scroll position, triggering additional scroll or resize events.
-   * Small deltas are ignored to prevent such feedback loops from causing
-   * oscillations at virtualization boundaries.
-   *
-   * Increase if you expect high inaccuracy in height measurements, or if you
-   * notice oscillations at virtualization boundaries.
-   *
-   * Default: 50px.
-   */
-  minScrollAndResizeDeltaPx?: number;
-
-  /**
-   * "Column editing mode" is a mode in which controls that are normally hidden
-   * become visible, such as column reorder and delete buttons.
-   *
-   * When this prop is undefined, the component manages the mode internally.
-   * Users can toggle it by tapping the table header on mobile or by clicking
-   * the column edit button, if enabled.
-   *
-   * Pass `true` or `false` to control the mode externally.
-   */
-  columnEditing?: boolean;
-
-  /**
-   * Called when the user requests to enter or leave column editing mode.
-   *
-   * The `source` parameter indicates what triggered the request.
-   *
-   * When `columnEditing` is not controlled, the component automatically
-   * applies the requested change internally.
-   *
-   * When `columnEditing` is controlled externally and you still want to
-   * respond to user requests, use this callback to decide whether to
-   * update the mode.
-   */
-  onColumnEditingRequest?: (editing: boolean, source: 'header' | 'edit-button') => void;
-
-  /**
-   * Whether to show a small gear button in the top-right corner that
-   * toggles column editing mode.
-   */
-  columnEditButton?: boolean;
-
-  /**
-   * Applied to the `<thead>` element.
-   */
-  theadClassName?: string;
-
-  /**
-   * Applied to the only `<tr>` element within the `<thead>`.
-   */
-  theadTrClassName?: string;
-
-  /**
-   * Applied to the `<tbody>` element.
-   */
-  tbodyClassName?: string;
-}
-
-export type SortOrder = Extract<AriaAttributes['aria-sort'], 'none' | 'ascending' | 'descending'>;
+import styles from './table.css';
 
 /**
- * The column specification.
+ * The new Table component. Use it instead of tables in the `legacy-table` folder.
+ *
+ * For every prop and component referenced here, see the corresponding docs
+ * for detailed behavior.
+ *
+ * ## Minimal usage
+ *
+ * You need the following props:
+ * - `data`
+ * - `getKey`
+ * - `columns`
+ *   - `key`
+ *   - `name` (not required but needed in most cases)
+ *   - `renderCell` (not required but needed in most cases)
+ *
+ * ## Selection
+ *
+ * Selection is handled on item level via `renderItem` (often with
+ * `DefaultItemRenderer`) and its props:
+ *
+ * - `clickable`
+ * - `selected`
+ * - `onClick` or `onPointerUp`, etc.
+ *
+ * You may use the TableSelection from `global/table-selection.ts` for the selection control:
+ *
+ * - `selected={tableSelection.isSelected(item)}`
+ * - `onClick={() => setTableSelection(tableSelection.toggle(item))}`
+ *
+ * See the stories for examples with this utility.
+ *
+ * Additionally, for accessibility, you will likely need a cell with a checkbox
+ * to toggle item selection.
+ *
+ * ## Sorting
+ *
+ * You need the following to support sorting:
+ *
+ * - Include `<SortButton />` in a column header
+ * - Set initial `Column.sortOrder` to `none`. Do not leave `undefined`
+ *   for accessibility reasons.
+ * - Handle `TableProps.onSort` callback in the client code. It is expected
+ *   to update `columns` by setting the new `sortOrder` value for
+ *   the corresponding column, and updating the data accordingly.
+ *
+ * ## Focus
+ *
+ * The table supports the ["roving tabindex"](https://developer.mozilla.org/en-US/docs/Web/Accessibility/Guides/Keyboard-navigable_JavaScript_widgets#technique_1_roving_tabindex)
+ * technique to focus rows with the up/down arrow keys and, possibly, with the pointer.
+ * To support it, use the following props of the `DefaultItemRenderer`:
+ *
+ * - `keyboardFocusable`
+ * - Possibly `onClick` invoking `focusRow(e.currentTarget)`
+ *
+ * In your custom row renderer, use `TableRow`, which also has the `keyboardFocusable` prop.
+ *
+ * ## Deleting columns
+ *
+ * You need the following to support deleting columns:
+ *
+ * - Make sure the `column` has a proper `name` or `key` prop, which will be
+ *   automatically included in the aria-label of `<DeleteColumnButton />`.
+ * - Include `<DeleteColumnButton />` in a column header
+ * - Handle `TableProps.onColumnDelete` callback in the client code. It is expected
+ *   to update `columns` by removing the corresponding column.
+ *
+ * ## Row virtualization
+ *
+ * To render only rows near the viewport and replace others with spacers, use:
+ *
+ * - `virtualizeRows`
+ * - `scrollerRef` — required when the scrollable container is not the whole document
+ * - `estimateHeight` — recommended when rows are expected to be taller than
+ *   the default height (e.g. multiline or custom content)
+ * - Fine-tuning props: `lookaheadPx`, `retentionMarginPx`, `minScrollAndResizeDeltaPx`
  */
-export interface Column<T> {
-  /**
-   * Used to render a row key, e.g. `<thead><tr><td key={getKey(item, index, items)}...</td></tr></thead>`.
-   */
-  key: React.Key;
+export default function Table<T>(props: TableProps<T> & ComponentPropsWithRef<'table'>) {
+  const {
+    data,
+    columns,
+    getKey,
+    noHeader,
+    onItemMove,
+    onSort,
+    onColumnDelete,
+    onColumnReorder,
+    noColumnReorderAnimation,
+    renderItem,
+    virtualizeRows = false,
+    scrollerRef,
+    estimateHeight = () => defaultRowHeight,
+    // eslint-disable-next-line no-magic-numbers
+    lookaheadPx = 400,
+    // eslint-disable-next-line no-magic-numbers
+    retentionMarginPx = 450,
+    // eslint-disable-next-line no-magic-numbers
+    minScrollAndResizeDeltaPx = 50,
+    columnEditing,
+    onColumnEditingRequest,
+    columnEditButton,
+    theadClassName,
+    theadTrClassName,
+    tbodyClassName,
 
-  /**
-   * Used in aria-labels of controls which do not contain text,
-   * e.g. `DeleteColumnButton`. If not set, the `String(key)` is used.
-   */
-  name?: string;
+    ref: userRef,
+    className,
+    ...restProps
+  } = props;
 
-  /**
-   * Default: name ?? String(key)
-   */
-  renderHeader?: () => ReactNode;
+  const localRef = useRef<HTMLTableElement>(null);
 
-  /**
-   * Renders a single cell value for a column.
-   * Default:
-   * - If item is an Array, renders `String(item[columnIndex])`
-   * - If item is an Object, renders `Object.values(item)[columnIndex]`
-   * - Otherwise:
-   *   - The first column renders `String(item)`
-   *   - Other columns render empty value
-   */
-  renderCell?: (item: T, index: number, items: T[]) => ReactNode;
+  const {virtualItems, intersectionObserverHandle, collapseItemIntoSpacer} = useVirtualItems({
+    enabled: virtualizeRows,
+    data,
+    scrollerRef,
+    tableRef: localRef,
+    estimateHeight,
+    lookaheadPx,
+    retentionMarginPx,
+    minScrollAndResizeDeltaPx,
+  });
 
-  /**
-   * If the column gets an indent when `DefaultItemRendererProps.level` returns
-   * a positive number.
-   */
-  indent?: boolean;
+  const [animatedColumn, setAnimatedColumn] = useState<AnimatedColumn | null>(null);
+  useAnimatedColumn({
+    animatedColumn,
+    setAnimatedColumn,
+    disabled: noColumnReorderAnimation,
+    tableRef: localRef,
+    columns,
+  });
 
-  /**
-   * If set, displays sort button and includes `aria-sort` in `th`.
-   * Handle clicks with {@link TableProps.onSort}.
-   */
-  sortOrder?: AriaAttributes['aria-sort'];
+  const handleRowNavigation = useCallback((e: React.KeyboardEvent<HTMLTableSectionElement>) => {
+    const arrowUp = e.key === 'ArrowUp';
+    const arrowDown = e.key === 'ArrowDown';
+    if (!arrowUp && !arrowDown) return;
 
-  /**
-   * Whether to display a delete button in the column header.
-   * Handle delete requests with {@link TableProps.onColumnDelete}.
-   */
-  deletable?: boolean;
+    const currentRow = (e.target as HTMLElement).closest('tr');
+    if (currentRow?.parentElement?.parentElement !== localRef.current) {
+      return;
+    }
 
-  /**
-   * Displays a reorder handle in the column header.
-   * Handle reorder requests with {@link TableProps.onColumnReorder}.
-   * If a function is provided, it determines whether the column may be moved
-   * to the specified insertion position.
-   */
-  canReorder?: boolean | ((insertionIndex: number, columns: Column<T>[]) => boolean);
+    let candidate: HTMLTableRowElement | null = currentRow;
+    while (candidate) {
+      candidate = (
+        arrowUp ? candidate.previousElementSibling : candidate.nextElementSibling
+      ) as HTMLTableRowElement | null;
 
-  /**
-   * The classname to apply to the `th` element inside `table / thead`.
-   */
-  thClassName?: string;
+      if (candidate?.hasAttribute(keyboardFocusableAttrName)) {
+        focusWithTemporaryTabIndex(candidate);
+        e.preventDefault();
+        return;
+      }
+    }
+  }, []);
 
-  /**
-   * The classname to apply to the `td` element inside `table / tbody`.
-   * If a custom `TableProps.renderItem` is provided, this prop is not used,
-   * unless the custom renderer falls back to `DefaultItemRenderer`.
-   */
-  tdClassName?: string | ((item: T, index: number, items: T[]) => string | undefined);
+  return (
+    <TablePropsContext value={props as TableProps<unknown>}>
+      <AnimatedColumnContext value={animatedColumn}>
+        <table className={classNames(styles.table, className)} ref={useComposedRef(userRef, localRef)} {...restProps}>
+          <TableHeader />
+          <IntersectionObserverContext value={intersectionObserverHandle}>
+            {/* eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions */}
+            <tbody className={tbodyClassName} onKeyDown={handleRowNavigation}>
+              {virtualItems.map(virtualItem => {
+                if (virtualItem.type === 'spacer') {
+                  return <SpacerRow key={virtualItem.key} spacer={virtualItem} colSpan={columns.length} />;
+                }
+
+                const index = virtualItem.index;
+                if (index < 0 || index >= data.length) return null;
+
+                const item = data[index];
+                const key = getKey(item, index, data);
+                return (
+                  <CollapseItemIntoSpacerContext value={height => collapseItemIntoSpacer(index, height)} key={key}>
+                    {renderItem ? renderItem(item, index, data) : <DefaultItemRenderer index={index} />}
+                  </CollapseItemIntoSpacerContext>
+                );
+              })}
+            </tbody>
+          </IntersectionObserverContext>
+        </table>
+      </AnimatedColumnContext>
+    </TablePropsContext>
+  );
 }
